fastapi_service/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
bcrypt==3.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.13.4
google-search-results==2.4.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
kafka==1.3.5
kafka-python==2.0.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
passlib==1.7.4
platformdirs==4.2.1
pluggy==1.5.0
pyarrow==16.0.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.0
pydantic-core==2.18.1
PyJWT==2.8.0
pymongo==4.6.3
pyOpenSSL==24.1.0
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.9.1
sortedcontainers==2.4.0
starlette==0.37.2
tomli==2.0.1
tomlkit==0.12.4
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
uvicorn==0.29.0
